[OnlyConnect] Add support for media questions
Users can now set a question type of 'media' for questions in the Connections
and Sequences rounds. This takes the form of the clue being replaced with a blob,
that gets rendered as an image.

The code currently does not enforce that all the clues in a question are either
text or media; it is technically possible to create mixed questions via a specific
sequence of inputs on the CMS